backend/pyproject.toml
[project]
name = "nova-rag-engine"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.1.0",
    "langchain-text-splitters>=0.0.2",
    "pypdf>=3.17.0",
    "pdfplumber>=0.10.3",
    "pymupdf>=1.24.0",
    "python-docx>=1.1.0",
    "python-pptx>=0.6.21",
    "lxml>=4.9.0",
    "pandas>=2.0.0",
    "openpyxl>=3.1.0",
    "tabulate>=0.9.0",
    "markdown>=3.5.2",
    "pgvector>=0.2.0",
    "psycopg2-binary>=2.9.0",
    "openai>=1.0.0",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "python-multipart>=0.0.9",
    "httpx>=0.27.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "sqlalchemy>=2.0.0",
    "rank-bm25>=0.2.2",
    "jieba>=0.42.1",
    "tiktoken>=0.7.0",
    "anyio>=4.13.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
